+++ packages/components/pie-cookie-banner/src/defs.ts
@@ -86,8 +86,8 @@ export type PreferenceIds = 'all' | 'necessary' | 'functional' | 'analytical' | 
 
 export interface Preference {
     id: PreferenceIds;
-    isDisabled?: boolean;
-    isChecked?: boolean;
+    checked?: boolean;
+    disabled?: boolean;
     hasDivider?: boolean;
     hasDescription?: boolean;
 }
@@ -100,8 +100,8 @@ export const preferences: Preference[] = [
     },
     {
         id: 'necessary',
-        isDisabled: true,
-        isChecked: true,
+        checked: true,
+        disabled: true,
         hasDescription: true,
     },
     {