+++ packages/components/pie-switch/src/defs.ts
@@ -15,11 +15,15 @@ export interface SwitchProps {
     /**
      * Same as the HTML checked attribute - indicates whether the switch is on or off
      */
-    isChecked?: boolean;
+    checked?: boolean;
     /**
-     * Same as the HTML checked attribute - indicates whether the switch disabled or not
+     * Same as the HTML required attribute - indicates whether the switch must be turned or not
      */
-    isDisabled?: boolean;
+    required: boolean;
+    /**
+     * Same as the HTML disabled attribute - indicates whether the switch is disabled or not
+     */
+    disabled?: boolean;
     /**
      * The label value of the component
      */
@@ -28,6 +32,14 @@ export interface SwitchProps {
      * The placement of the label such as leading or trailing
      */
     labelPlacement?: LabelPlacement;
+    /**
+     * Same as the HTML name attribute - indicates the name of the switch (for use with forms)
+     */
+    name?: string;
+    /**
+     * Same as the HTML value attribute - indicates the value of the switch (for use with forms). Defaults to 'on'.
+     */
+    value: string;
 }
 
 /**